+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### 23.07.2018
 
+* [Rust] Code of Kutulu - Fixed unwrap on closest explorer (CodinGame)
 * [Rust] Code of Kutulu - Fixed running away (CodinGame)
 
 #### 22.07.2018